Fix speaker buffer size
The buffer size was previously set on the speaker but during the upgrade of the Oto version, was in error moved to the context. These aren't the same and this change moves setting the buffer size back to the player. Commit introducing the bug: https://github.com/gopxl/beep/commit/5fefa491fdb0ba5eba5854f06fbd3cce91f4f9f2